# only calculate a document's CRC32 if needed