--- a/crengine/src/lvdocview.cpp
+++ b/crengine/src/lvdocview.cpp
@@ -4117,7 +4117,8 @@ bool LVDocView::LoadDocument(const lChar32 * fname, bool metadataOnly) {
 		m_doc_props->setString(DOC_PROP_FILE_SIZE, lString32::itoa(
 				(int) stream->GetSize()));
 		m_doc_props->setString(DOC_PROP_FILE_NAME, arcItemPathName);
-        m_doc_props->setHex(DOC_PROP_FILE_CRC32, stream->getcrc32());
+		if (!metadataOnly)
+			m_doc_props->setHex(DOC_PROP_FILE_CRC32, stream->getcrc32());
 		// loading document
 		if (LoadDocument(stream, metadataOnly)) {
 			m_filename = lString32(fname);
@@ -4163,7 +4164,8 @@ bool LVDocView::LoadDocument(const lChar32 * fname, bool metadataOnly) {
     m_doc_props->setString(DOC_PROP_FILE_NAME, fn);
 	m_doc_props->setString(DOC_PROP_FILE_SIZE, lString32::itoa(
 			(int) stream->GetSize()));
-    m_doc_props->setHex(DOC_PROP_FILE_CRC32, stream->getcrc32());
+	if (!metadataOnly)
+		m_doc_props->setHex(DOC_PROP_FILE_CRC32, stream->getcrc32());
 
 	if (LoadDocument(stream, metadataOnly)) {
 		m_filename = lString32(fname);
@@ -4657,7 +4659,8 @@ bool LVDocView::LoadDocument(LVStreamRef stream, bool metadataOnly) {
 					m_doc_props->setString(DOC_PROP_FILE_NAME, fn);
 					m_doc_props->setString(DOC_PROP_CODE_BASE, LVExtractPath(fn) );
 					m_doc_props->setString(DOC_PROP_FILE_SIZE, lString32::itoa((int)m_stream->GetSize()));
-                    m_doc_props->setHex(DOC_PROP_FILE_CRC32, m_stream->getcrc32());
+					if (!metadataOnly)
+						m_doc_props->setHex(DOC_PROP_FILE_CRC32, m_stream->getcrc32());
 					found = true;
 				}
 			}