It's only used by the caching, so only calculate it on full document load (i.e. not metadata only).

Resulting speedup during metadata extraction:
- simulator on my laptop (1068 files): ~9% speedup, average ~12 → ~10 ms, minimum 5 → 5 ms, maximum 486 → 319 ms
- on my Kindle Paperwhite (6th generation, 956 files): ×2 speedup: average 225 → 107 ms, minimum 84 → 84 ms, maximum 4862 → 400 ms

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/koreader/crengine/498)
<!-- Reviewable:end -->